fix: creating uuids we do not intent to store
We should really not create UUIDs if we don't intent to store them for
anywhere. So in favor of keeping this a stateless program we default to
the transaction ID from Nordigen and keep it a string in Ynabber. If the
ID from Nordigen is empty we warn the user in the logs.