feat(nordigen)!: add requisition hook
Rework the requisition fulfillment to be under the Reader struct and
simplify the file handling.

Move the requisition hook to the Nordigen config struct and make it more
generic so i can be used to handle arbitrary events tied to the
requisition and its status.

BREAKING CHANGE: Remove NORDIGEN_DATAFILE in favor of the
YNABBER_DATADIR option. This will require recreating the requisition.