feat(nordigen): add Nordea specific mapper
Add mapper thats specific to Nordea which has some quirks that needs to
be handled differently than the default mapper.

I would like to get rid of the logic inside the default mapper
eventually and have all the logic in bank specific mappers. The goal of
this is to have less config values and less stuff the user needs to
reason about. Its getting hard to remember how setting X affects each
bank and what it needs to be for each bank, so lets just put it directly
in code.