fix(nordigen): nil pointer
Refactor toYnabber into a single function, the nil check was on a layer
too high before. Remove that pit fall by just having one function for
the translation.

Move the strip function into the mapper, it was a leftover from when the
config value was defined on the YNAB side. We want to mangle data closet
to the read point.

commit-id:ca48367d