fix: not crashing on error
If we get an error in the reader, close it and let the program crash. 
Handle retires if rate limited as long as we are not running in on-shot 
mode aka interval=0.

If a writer fails or we have multiple readers and not all of them 
fail, the program could hang in a suboptimal state. We need to solve 
that eventually.

Close: #117
Fix: #110