fix(nordigen): No Payee Set for Nordea
For some reason Nordea decided to change its behaivor again. It now 
started providing its remittance info in the sturcutred version. I’m 
taking a gamble here and included it in the remittance source. There is 
no way for me to know if this will work for everyone but it should be 
sane and only use it if its defined.

I’m renmaing from unstructred to remittance across the board but keeps 
backwards compatibility doing config init.